requirements.txt
openai>=1.99.0
python-dotenv>=1.0.1
pytest>=8.0.0
